pyproject.toml
[project]
name = "graph-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sage", email = "sage.b.shaw@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ipython>=9.8.0",
]
